ASoC: ymu831: Fix overflow when casting int to uint8
The value returned by get_incall_mic() is always used as
int and it may be -ERRNO, so it should not be returned as uint8:

sound/soc/codecs/ymu831/ymu831.c:1652 get_incall_mic() warn: signedness bug returning '(-5)'

+++ sound/soc/codecs/ymu831/ymu831.c
@@ -1609,7 +1609,7 @@ static int get_unused_mic_block_on(void)
 }
 
 
-static UINT8 get_incall_mic(
+static int get_incall_mic(
 	struct snd_soc_codec *codec,
 	int	output_path
 )
